+++ app/assets/stylesheets/pages/_home.scss
@@ -1,6 +1,6 @@
 // Specific CSS for your home-page
 .container {
-  margin-top: 100px;
+  margin-top: 50px;
   display: flex;
   flex-direction: column;
   justify-content: left;
