--- app/controllers/establishment_controller.rb
@@ -1,22 +1,0 @@
-class EstablishmentController < ApplicationController
- def index
-   @user = current_user
-   @establishments = Establishment.all
- end
-
- def show
-  @user = current_user
-  @establishment = Establishment.find(params[:id])
- end
-
- def new
-  @user = current_user
-  @establishment = Establishment.new
-end
-
- private
-
- def establishment_params
-  params.require(:establishment).permit(:name, :phone, :address, :type, :rating, :image, :photo)
- end
-end
