+++ app/controllers/establishments_controller.rb
@@ -1,0 +1,27 @@
+class EstablishmentsController < ApplicationController
+  def index
+    @user = current_user
+    @establishments = Establishment.all
+  end
+
+  def show
+    @user = current_user
+    @establishment = Establishment.find(params[:id])
+  end
+
+  def new
+    @user = current_user
+    @establishment = Establishment.new
+
+  def create
+
+  end
+
+end
+
+ private
+
+ def establishment_params
+  params.require(:establishment).permit(:name, :phone, :address, :type, :rating, :image, :photo)
+ end
+end
