+++ app/javascript/controllers/collapsable_menu_controller.js
@@ -1,0 +1,16 @@
+import { Controller } from "@hotwired/stimulus"
+
+export default class extends Controller {
+  MENU = document.getElementById("menu_logo")
+
+  MENU.addEventListener("click", (event) => {
+      if ($("#navbar").hasClass("clickHide")) {
+        $("#navbar").removeClass("clickHide");
+        $("#menu_logo").html("-");
+      }
+      else {
+        $("#navbar").addClass("clickHide");
+        $("#menu_logo").html("+");
+      }
+  });
+}
