+++ app/views/establishments/_new_form.html.erb
@@ -1,0 +1,27 @@
+<div class="container">
+    <%= form_with model: @establishment do |f| %>
+      <div class="mb-3">
+        <%= f.label :name, {class: 'form-label'} %>
+        <%= f.text_field :name, {class: 'form-control'} %>
+      </div>
+      <div class="mb-3">
+        <%= f.label :address, {class: 'form-label'} %>
+        <%= f.text_field :address, {class: 'form-control'} %>
+      </div>
+      <div class="mb-3">
+        <%= f.label :rating, {class: 'form-label'} %>
+        <%= f.number_field :rating, {class: 'form-control'} %>
+      </div>
+      <div class="mb-3">
+        <%= f.label :phone, {class: 'form-label'} %>
+        <%= f.number_field :phone, {class: 'form-control'} %>
+      </div>
+      <div class="mb-3">
+        <%= f.label :type, "Location-type", {class: 'form-label'} %>
+        <%= f.collection_select :type, Establishment.types.keys, :to_s, :to_s, prompt: "Select a type" %>
+      </div>
+      <div class="mb-3">
+        <%= f.submit "Submit" %>
+      </div>
+    <% end %>
+ </div>
