+++ app/views/establishments/_searchbar.html.erb
@@ -1,0 +1,14 @@
+  <div class="container">
+    <form novalidate="novalidate" class="simple_form search" action="/" accept-charset="UTF-8" method="get">
+    <div class="search-form-control form-group">
+      <input class="form-control string required" type="text" name="search[query]" id="search_query" />
+      <input
+        name="address" placeholder="Address" type="text"
+        autocomplete="address-line1"
+      />
+      <button name="button" type="submit" class="btn btn-flat">
+        <i class="fas fa-search"></i> Search
+      </button>
+    </div>
+  </form>
+</div>
