+++ app/views/shared/_navbar.html.erb
@@ -5,8 +5,8 @@
     </div>
     <nav>
       <ul>
-        <li><%= link_to "My Kennel", "#" %></li>
         <li><%= link_to "My Pawfect List", "#" %></li>
+        <li><%= link_to "+ dog friendly location", new_establishment_path %></li>
         <li><%= link_to "Log out", destroy_user_session_path, data: {turbo_method: :delete} %></li>
       </ul>
     </nav>
@@ -15,5 +15,4 @@
       <%= link_to "Login", new_user_session_path, class: "nav-link" %>
     </li>
   <% end %>
-  <script src="/app/javascript/controllers/collapsable_menu.js"></script>
 </div>
