+++ db/migrate/20230105093049_change_establishment_type_to_integer.rb
@@ -1,0 +1,7 @@
+class ChangeEstablishmentTypeToInteger < ActiveRecord::Migration[7.0]
+
+  def change
+    remove_column :establishments, :type, :string
+    add_column :establishments, :type, :integer, default: 0
+  end
+end
